refactor: Meeting 더미, 타입 분리
